+++ langs/python/src/vmx/collections/__init__.py
@@ -22,6 +22,7 @@ from vmx.collections.collection_changed import CollectionChangedEvent
 from vmx.collections.observable_dictionary import ObservableDictionary
 from vmx.collections.observable_list import ObservableList
 from vmx.collections.paged_composition import PagedComposition
+from vmx.collections.protocols import SelectableVmCollectionProto, VmCollectionProto
 from vmx.collections.serviced_observable_collection import ServicedObservableCollection
 from vmx.collections.token_paged_composition import TokenPagedComposition
 
@@ -31,6 +32,8 @@ __all__ = [
     "ObservableDictionary",
     "ObservableList",
     "PagedComposition",
+    "SelectableVmCollectionProto",
     "ServicedObservableCollection",
     "TokenPagedComposition",
+    "VmCollectionProto",
 ]